fix: move HTTP out of tweaks dir
