+++ app/src/main/kotlin/io/github/revenge/xposed/HTTP.kt
@@ -1,6 +1,5 @@
-package io.github.revenge.xposed.tweaks
+package io.github.revenge.xposed
 
-import io.github.revenge.xposed.RevengeConstants
 import io.ktor.client.*
 import io.ktor.client.call.*
 import io.ktor.client.engine.cio.*
